updat

+++ atom/browser/browser_win.cc
@@ -4,12 +4,15 @@
 
 #include "atom/browser/browser.h"
 
+#include <windows.h>  // windows.h must be included first
+
 #include <atlbase.h>
-#include <propkey.h>
-#include <windows.h>
 #include <shlobj.h>
 #include <shobjidl.h>
 
+#include "atom/browser/ui/win/jump_list.h"
+#include "atom/common/atom_version.h"
+#include "atom/common/native_mate_converters/string16_converter.h"
 #include "base/base_paths.h"
 #include "base/file_version_info.h"
 #include "base/files/file_path.h"
@@ -17,10 +20,9 @@
 #include "base/strings/string_util.h"
 #include "base/strings/stringprintf.h"
 #include "base/strings/utf_string_conversions.h"
-#include "base/win/win_util.h"
 #include "base/win/registry.h"
+#include "base/win/win_util.h"
 #include "base/win/windows_version.h"
-#include "atom/common/atom_version.h"
 
 namespace atom {
 
@@ -39,6 +41,28 @@ BOOL CALLBACK WindowsEnumerationHandler(HWND hwnd, LPARAM param) {
   }
 
   return TRUE;
+}
+
+bool GetProtocolLaunchPath(mate::Arguments* args, base::string16* exe) {
+  // Executable Path
+  if (!args->GetNext(exe)) {
+    base::FilePath path;
+    if (!PathService::Get(base::FILE_EXE, &path)) {
+      LOG(ERROR) << "Error getting app exe path";
+      return false;
+    }
+    *exe = path.value();
+  }
+
+  // Read in optional args arg
+  std::vector<base::string16> launch_args;
+  if (args->GetNext(&launch_args) && !launch_args.empty())
+    *exe = base::StringPrintf(L"\"%s\" %s \"%%1\"",
+                              exe->c_str(),
+                              base::JoinString(launch_args, L" ").c_str());
+  else
+    *exe = base::StringPrintf(L"\"%s\" \"%%1\"", exe->c_str());
+  return true;
 }
 
 }  // namespace
@@ -79,70 +103,41 @@ void Browser::SetAppUserModelID(const base::string16& name) {
   SetCurrentProcessExplicitAppUserModelID(app_user_model_id_.c_str());
 }
 
-void Browser::SetUserTasks(const std::vector<UserTask>& tasks) {
-  CComPtr<ICustomDestinationList> destinations;
-  if (FAILED(destinations.CoCreateInstance(CLSID_DestinationList)))
-    return;
-  if (FAILED(destinations->SetAppID(GetAppUserModelID())))
-    return;
+bool Browser::SetUserTasks(const std::vector<UserTask>& tasks) {
+  JumpList jump_list(GetAppUserModelID());
+  if (!jump_list.Begin())
+    return false;
 
-  // Start a transaction that updates the JumpList of this application.
-  UINT max_slots;
-  CComPtr<IObjectArray> removed;
-  if (FAILED(destinations->BeginList(&max_slots, IID_PPV_ARGS(&removed))))
-    return;
-
-  CComPtr<IObjectCollection> collection;
-  if (FAILED(collection.CoCreateInstance(CLSID_EnumerableObjectCollection)))
-    return;
-
-  for (auto& task : tasks) {
-    CComPtr<IShellLink> link;
-    if (FAILED(link.CoCreateInstance(CLSID_ShellLink)) ||
-        FAILED(link->SetPath(task.program.value().c_str())) ||
-        FAILED(link->SetArguments(task.arguments.c_str())) ||
-        FAILED(link->SetDescription(task.description.c_str())))
-      return;
-
-    if (!task.icon_path.empty() &&
-        FAILED(link->SetIconLocation(task.icon_path.value().c_str(),
-                                     task.icon_index)))
-      return;
-
-    CComQIPtr<IPropertyStore> property_store = link;
-    if (!base::win::SetStringValueForPropertyStore(property_store, PKEY_Title,
-                                                   task.title.c_str()))
-      return;
-
-    if (FAILED(collection->AddObject(link)))
-      return;
+  JumpListCategory category;
+  category.type = JumpListCategory::Type::TASKS;
+  category.items.reserve(tasks.size());
+  JumpListItem item;
+  item.type = JumpListItem::Type::TASK;
+  for (const auto& task : tasks) {
+    item.title = task.title;
+    item.path = task.program;
+    item.arguments = task.arguments;
+    item.icon_path = task.icon_path;
+    item.icon_index = task.icon_index;
+    item.description = task.description;
+    category.items.push_back(item);
   }
 
-  // When the list is empty "AddUserTasks" could fail, so we don't check return
-  // value for it.
-  CComQIPtr<IObjectArray> task_array = collection;
-  destinations->AddUserTasks(task_array);
-  destinations->CommitList();
+  jump_list.AppendCategory(category);
+  return jump_list.Commit();
 }
 
-bool Browser::RemoveAsDefaultProtocolClient(const std::string& protocol) {
+bool Browser::RemoveAsDefaultProtocolClient(const std::string& protocol,
+                                            mate::Arguments* args) {
   if (protocol.empty())
     return false;
 
-  base::FilePath path;
-  if (!PathService::Get(base::FILE_EXE, &path)) {
-    LOG(ERROR) << "Error getting app exe path";
-    return false;
-  }
-
   // Main Registry Key
   HKEY root = HKEY_CURRENT_USER;
-  std::string keyPathStr = "Software\\Classes\\" + protocol;
-  std::wstring keyPath = std::wstring(keyPathStr.begin(), keyPathStr.end());
+  base::string16 keyPath = base::UTF8ToUTF16("Software\\Classes\\" + protocol);
 
   // Command Key
-  std::string cmdPathStr = keyPathStr + "\\shell\\open\\command";
-  std::wstring cmdPath = std::wstring(cmdPathStr.begin(), cmdPathStr.end());
+  base::string16 cmdPath = keyPath + L"\\shell\\open\\command";
 
   base::win::RegKey key;
   base::win::RegKey commandKey;
@@ -154,13 +149,15 @@ bool Browser::RemoveAsDefaultProtocolClient(const std::string& protocol) {
     // Key doesn't even exist, we can confirm that it is not set
     return true;
 
-  std::wstring keyVal;
+  base::string16 keyVal;
   if (FAILED(commandKey.ReadValue(L"", &keyVal)))
     // Default value not set, we can confirm that it is not set
     return true;
 
-  std::wstring exePath(path.value());
-  std::wstring exe = L"\"" + exePath + L"\" \"%1\"";
+  base::string16 exe;
+  if (!GetProtocolLaunchPath(args, &exe))
+    return false;
+
   if (keyVal == exe) {
     // Let's kill the key
     if (FAILED(key.DeleteKey(L"shell")))
@@ -172,7 +169,8 @@ bool Browser::RemoveAsDefaultProtocolClient(const std::string& protocol) {
   }
 }
 
-bool Browser::SetAsDefaultProtocolClient(const std::string& protocol) {
+bool Browser::SetAsDefaultProtocolClient(const std::string& protocol,
+                                        mate::Arguments* args) {
   // HKEY_CLASSES_ROOT
   //    $PROTOCOL
   //       (Default) = "URL:$NAME"
@@ -190,26 +188,17 @@ bool Browser::SetAsDefaultProtocolClient(const std::string& protocol) {
   if (protocol.empty())
     return false;
 
-  base::FilePath path;
-  if (!PathService::Get(base::FILE_EXE, &path)) {
-    LOG(ERROR) << "Error getting app exe path";
+  base::string16 exe;
+  if (!GetProtocolLaunchPath(args, &exe))
     return false;
-  }
 
   // Main Registry Key
   HKEY root = HKEY_CURRENT_USER;
-  std::string keyPathStr = "Software\\Classes\\" + protocol;
-  std::wstring keyPath = std::wstring(keyPathStr.begin(), keyPathStr.end());
-  std::string urlDeclStr = "URL:" + protocol;
-  std::wstring urlDecl = std::wstring(urlDeclStr.begin(), urlDeclStr.end());
+  base::string16 keyPath = base::UTF8ToUTF16("Software\\Classes\\" + protocol);
+  base::string16 urlDecl = base::UTF8ToUTF16("URL:" + protocol);
 
   // Command Key
-  std::string cmdPathStr = keyPathStr + "\\shell\\open\\command";
-  std::wstring cmdPath = std::wstring(cmdPathStr.begin(), cmdPathStr.end());
-
-  // Executable Path
-  std::wstring exePath(path.value());
-  std::wstring exe = L"\"" + exePath + L"\" \"%1\"";
+  base::string16 cmdPath = keyPath + L"\\shell\\open\\command";
 
   // Write information to registry
   base::win::RegKey key(root, keyPath.c_str(), KEY_ALL_ACCESS);
@@ -224,24 +213,21 @@ bool Browser::SetAsDefaultProtocolClient(const std::string& protocol) {
   return true;
 }
 
-bool Browser::IsDefaultProtocolClient(const std::string& protocol) {
+bool Browser::IsDefaultProtocolClient(const std::string& protocol,
+                                      mate::Arguments* args) {
   if (protocol.empty())
     return false;
 
-  base::FilePath path;
-  if (!PathService::Get(base::FILE_EXE, &path)) {
-    LOG(ERROR) << "Error getting app exe path";
+  base::string16 exe;
+  if (!GetProtocolLaunchPath(args, &exe))
     return false;
-  }
 
   // Main Registry Key
   HKEY root = HKEY_CURRENT_USER;
-  std::string keyPathStr = "Software\\Classes\\" + protocol;
-  std::wstring keyPath = std::wstring(keyPathStr.begin(), keyPathStr.end());
+  base::string16 keyPath = base::UTF8ToUTF16("Software\\Classes\\" + protocol);
 
   // Command Key
-  std::string cmdPathStr = keyPathStr + "\\shell\\open\\command";
-  std::wstring cmdPath = std::wstring(cmdPathStr.begin(), cmdPathStr.end());
+  base::string16 cmdPath = keyPath + L"\\shell\\open\\command";
 
   base::win::RegKey key;
   base::win::RegKey commandKey;
@@ -253,19 +239,13 @@ bool Browser::IsDefaultProtocolClient(const std::string& protocol) {
     // Key doesn't exist, we can confirm that it is not set
     return false;
 
-  std::wstring keyVal;
+  base::string16 keyVal;
   if (FAILED(commandKey.ReadValue(L"", &keyVal)))
     // Default value not set, we can confirm that it is not set
     return false;
 
-  std::wstring exePath(path.value());
-  std::wstring exe = L"\"" + exePath + L"\" \"%1\"";
-  if (keyVal == exe) {
-    // Default value is the same as current file path
-    return true;
-  } else {
-    return false;
-  }
+  // Default value is the same as current file path
+  return keyVal == exe;
 }
 
 bool Browser::SetBadgeCount(int count) {
@@ -273,13 +253,13 @@ bool Browser::SetBadgeCount(int count) {
 }
 
 void Browser::SetLoginItemSettings(LoginItemSettings settings) {
-  std::wstring keyPath = L"Software\\Microsoft\\Windows\\CurrentVersion\\Run";
+  base::string16 keyPath = L"Software\\Microsoft\\Windows\\CurrentVersion\\Run";
   base::win::RegKey key(HKEY_CURRENT_USER, keyPath.c_str(), KEY_ALL_ACCESS);
 
   if (settings.open_at_login) {
     base::FilePath path;
     if (PathService::Get(base::FILE_EXE, &path)) {
-      std::wstring exePath(path.value());
+      base::string16 exePath(path.value());
       key.WriteValue(GetAppUserModelID(), exePath.c_str());
     }
   } else {
@@ -289,14 +269,14 @@ void Browser::SetLoginItemSettings(LoginItemSettings settings) {
 
 Browser::LoginItemSettings Browser::GetLoginItemSettings() {
   LoginItemSettings settings;
-  std::wstring keyPath = L"Software\\Microsoft\\Windows\\CurrentVersion\\Run";
+  base::string16 keyPath = L"Software\\Microsoft\\Windows\\CurrentVersion\\Run";
   base::win::RegKey key(HKEY_CURRENT_USER, keyPath.c_str(), KEY_ALL_ACCESS);
-  std::wstring keyVal;
+  base::string16 keyVal;
 
   if (!FAILED(key.ReadValue(GetAppUserModelID(), &keyVal))) {
     base::FilePath path;
     if (PathService::Get(base::FILE_EXE, &path)) {
-      std::wstring exePath(path.value());
+      base::string16 exePath(path.value());
       settings.open_at_login = keyVal == exePath;
     }
   }
